이미지DB 등록 성공. Mapper에서 insert문에 useGeneratedKeys 추가해서 auto_increment값을 vo 파라미터에 자동으로 변환되게 하고 서비스에서 insert 오버라이드를 Long 타입으로 반환하도록 하고 리턴값에 VO.getAno() 를 추가, 컨트롤러에서 기존DTO를 먼저 INSERT한뒤 Long ano 선언하고 이미지DTO에 auto_increment로 생성된 ano를 set한 뒤 이미지insert 실행.

+++ project_DiB/src/main/java/org/zerock/project_dib/accommodation/dto/AccommodationImgDTO.java
@@ -13,7 +13,7 @@ public class AccommodationImgDTO {
 
     private String file_name;
 
-    private int ano;
+    private Long ano;
 
     private int ord;
